1. Complate assists deteramine ; 2. Fix mistake of weapons spread in diffent display resolution

+++ Assets/Resources/Script/System/S_Bullet.cs
@@ -34,7 +34,7 @@ public class S_Bullet : ComponentSystem {
 
                             var attack = _bullet.attack;
                             // Modify demage by different body part
-                            attack.demage = _bullet.attack.demage * AvatarTagsData.demageRate[hitTag];
+                            attack.demage = (int)(_bullet.attack.demage * AvatarTagsData.demageRate[hitTag]);
 
                             // Set if head shot
                             attack.headShot = AvatarTagsData.IsHead(hitTag);
